requirements.txt
discord.py
python-dotenv
Daug==2023.12.9.1
